Created a marg_vs_divorce_viz notebook
Started to plot two line plots to observe and compare the trends of the divorce rate in Maine and Margarine consumption over time. The first plot has year as the x-axis and margarine consumption (lbs per capita) as the y-axis. The second plot has year as the x-axis and divorce rate in Maine (per 1000) as the y-axis.